rename source files
